rename contact action -> upsertContact action
